requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
dotenv==0.9.9
drf-yasg==1.21.8
inflection==0.5.1
isort==6.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
